refactor(schemas): split models.py into per-domain schema package (#729)
Split the 2,256-line models.py monolith (155 top-level definitions — every
Pydantic wire schema for the API) into a `schemas/` package with one module
per domain: bible, security, assessment, predict, agent, eflomal, tokenizer,
affix, tfidf, pivot, training, plus a shared validators module.

models.py becomes a thin re-export shim (`from schemas import *`) so the
frozen v3 surface's historical `from models import X` imports keep working
unchanged (epic #842). Import sites are intentionally left untouched.

Pure refactor — no behavior change. Nothing was renamed, retyped, or
re-configured; the mixed orm_mode/from_attributes configs and the _v3
naming are preserved verbatim. Verified:
- AST-level source comparison of old vs new: all 155 top-level defs present
  exactly once, byte-identical source, 0 missing / 0 duplicated / 0 mismatched.
- `from models import X` public surface unchanged (same 155 schema names).
- OpenAPI v3 contract snapshot test passes (byte-identical wire contract).
- Full `pytest test/` suite green (1175 passed); black/isort/pre-commit pass.
